implement the splitting of the classes
